feat: change predict.py to work in lower memory settings

+++ src/predict.py
@@ -45,7 +45,7 @@ logger.addHandler(file_handler)
 
 torch.backends.cudnn.allow_tf32 = True
 torch.backends.cuda.matmul.allow_tf32 = True
-compute_type = "float16"  # change to "int8" if low on GPU mem (may reduce accuracy)
+compute_type = "int8"  # change to "int8" if low on GPU mem (may reduce accuracy)
 device = "cuda"
 whisper_arch = "./models/faster-whisper-large-v3"
